Bench: refactor
