+++ benchmark/Makefile
@@ -1,12 +1,15 @@
 
 #CXX = clang++
 
-CXXFLAGS = -std=c++11 -I../src -g -O2 -DNDEBUG
+CXXFLAGS = -std=c++11  -I../benchmark -I../src -g -O2 #-DNDEBUG
 LDFLAGS = -L../src -lntl -lgmpxx -lgmp
+
+OBJS = benchmark.o
 
 all: benchmark
 
 benchmark: benchmark.o ../src/libntl.a
 	$(CXX) -o benchmark benchmark.o $(LDFLAGS)
+
 clean:
-	$(RM) *.o benchmark
+	$(RM) $(OBJS) benchmark
